pyproject.toml
[tool.poetry]
name = "2024kh-parents-attendance"
version = "0.1.0"
description = ""
authors = ["lyg6772 <lyg6772@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
uvicorn = "^0.29.0"
jinja2 = "^3.1.3"
aiomysql = "^0.2.0"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.29"
pymysql = "^1.1.0"
pyjwt = "^2.8.0"
passlib = "^1.7.4"
python-dateutil = "^2.9.0.post0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
